Push marker interface gen. classes to the package level
